+++ client/client_urls.py
@@ -7,12 +7,16 @@ from client.views import *
 urlpatterns = [
     path('', views.client_main_page, name='client'),  # main client page
     path('profile', views.client_profile, name='client_profile'),
+
+    # Team Rome - Edit Forms URLs
     path('edit', views.client_edit_main, name='client_edit'),
     path('edit/skills', views.client_edit_skills, name='client_edit_skills'),
     path('edit/photo', views.client_edit_photo, name='client_edit_photo'),
     path('edit/cv', views.client_edit_cv, name='client_edit_cv'),
     path('edit/education', views.client_edit_education, name='client_edit_education'),
     path('edit/experience', views.client_edit_experience, name='client_edit_experience'),
+    path('edit/form_edu/', views.FormEducation.as_view(), name='form_edu'),
+    # Team Rome
 
     path('chat/', login_required(views.MessagesView.as_view()), name='contact_with_centre'),
     path(r'opinion/', opinion_list, name='opinion_list'),
@@ -23,7 +27,7 @@ urlpatterns = [
     path('login/', client_login),
     path('logout/', client_logout),
     path('tasks/', tasks, name='tasks_list'),
-    path('edit/form_edu/', views.FormEducation.as_view(), name='form_edu'),
+
     path(r'checktask/', views.checktask),
     path(r'checknotifications/', views.checknotifications),
     path(r'settings/', views.settings_menu, name='settings_menu'),
