+++ client/templates/client/edit_forms/client_edit_education.html
@@ -134,7 +134,7 @@
                 </div>
             </div>
         {% empty %}
-            {% include 'client/education_empty.html' %}
+            {% include 'client/edit_forms/education_empty.html' %}
         {% endfor %}
 
         <div class="text-center">
